% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load demo data.R
\name{get.demo.data}
\alias{get.demo.data}
\title{Load Credit Card Name Data}
\usage{
get.demo.data(limit = NA, ...)
}
\arguments{
\item{limit}{Provide the maximum number of records to return.  Defaults to NA, or all available records.}

}
\description{
This function returns a character, or factor, vector of examples names that might appear on credit cards.
}
\examples{
x <- get.demo.data

x <- get.demo.data(10)
}
\keyword{data}
\keyword{demo}
\keyword{example}
