% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse names.R
\name{parse.names}
\alias{parse.names}
\title{Parse Names on Credit Cards}
\usage{
parse.names(x, ...)
}
\arguments{
\item{x}{Provide the charater vector to parse.}

\item{...}{clean up}
}
\description{
This function returns a data frame of parse attributes based on an input of credit card names.
}
\examples{
x <- get.demo.data
y <- parse.names(x)
}
\keyword{parse}
